// src/components/forms/checkins/annual/types.ts
export type AnnualFormValues = {
    agencyName: string;
    agencyType: string;
    agencySizeCategory: string;
    primaryServiceJurisdiction: string;
    geographicCoverageArea: string;
    totalAuthorizedPositions: string;
    totalFundedPositions: string;
    minimumSafeStaffingLevel: string;
    specialtyUnitPositionsCount: string;
    supervisorToStaffRatio: string;
    standardShiftLengthHours: string;
    shiftScheduleType: string;
    minimumRestPeriodPolicyExists: string;
    activePeerSupportTeam: string;
    hasEmployeeAssistanceProgram: string;
    goal1PrimaryAnnualGoal: string;
    goal1TargetMetric: string;
    goal1Timeframe: string;
    goal2SecondaryAnnualGoal: string;
    goal2TargetMetric: string;
    goal2Timeframe: string;
};